requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.2.0
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.115.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
pillow==10.4.0
psycopg==3.2.13
psycopg-binary==3.2.13
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.30.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
